pyproject.toml
[tool.poetry]
name = "zarr-visualization"
version = "0.1.0"
description = "Visualization dashboard for zarr data"
authors = ["Tobias Ferreira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
xarray = "^2024.3.0"
pyzmq = "24"
jupyter = "^1.0.0"
lab = "^8.1"
jupyterlab = "^4.1.6"
requests = "^2.31.0"
tqdm = "^4.66.2"
dask = "^2024.4.1"
pooch = "^1.8.1"
netcdf4 = "^1.6.5"
scipy = "^1.13.0"
ipywidgets-bokeh = "^1.6.0"
voila = "^0.5.6"
zarr = "^2.17.2"
aiohttp = "^3.9.5"
python-dotenv = "^1.0.1"
cachetools = "^5.3.3"
blinker = "^1.7.0"
s3fs = "^2024.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
